Moved EXIF.py module into singleshot source (since I started modifying it).
Added flag to sscgi to enable cgitb or to use a more circumspect traceback
handler.

Revise and replace IPTC and EXIF parsing with version that are much faster
(but only bother reading tags we use).

Add population of the various camera/exposure properties of ImageItems.

Change uptodate check of data cache to go ahead and stat all directories to
give us a much better shot of detecting when the cache needs refreshing.